pyproject.toml
[project]
name = "tts-poc"
version = "0.1.0"
description = "TTS PoC: Chatterbox + OpenVoice V2 for regional BR-PT and AR-ES accents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chatterbox-tts>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "torchaudio>=2.0.0",
    "python-multipart>=0.0.9",
    "soundfile>=0.12.0",
    "ml-dtypes>=0.5.0",
    "onnx>=1.17.0,<1.18",
    "datasets>=4.8.3",
    "resemble-perth>=1.0.1",
    "setuptools>=82.0.1",
]

# OpenVoice V2 has conflicting deps (gradio 3.x, numpy 1.22).
# Install separately if needed: pip install git+https://github.com/myshell-ai/OpenVoice.git

[project.scripts]
tts-server = "src.server:main"

[dependency-groups]
dev = ["pytest>=8.0"]
